test: split test_auto_e2e into per-package test modules
