+++ Model/tests/test_auto_e2e.py
@@ -3,18 +3,7 @@ import torch
 import sys
 sys.path.append('..')
 
-from model_components.backbone import Backbone
-from model_components.feature_fusion import FeatureFusion
-from model_components.trajectory_planning import (
-    GRUPlanner,
-    FlowMatchingPlanner,
-    build_planner,
-    PLANNER_REGISTRY,
-)
-from model_components.future_state import FutureState
-from model_components.view_fusion import build_view_fusion, FUSION_REGISTRY
-from model_components.view_fusion.cross_attention_fusion import CrossAttentionViewFusion
-from model_components.view_fusion.bev_fusion import BEVViewFusion
+from model_components.trajectory_planning import GRUPlanner
 
 
 def make_inputs(batch_size, num_views, device, include_camera_params=False):
@@ -70,7 +59,7 @@ class TestBatchIndependence:
         traj_both, _, _ = model(visual, map_input, vis_hist, ego, mode="infer")
 
         # Single sample forward (sample 0)
-        traj_single, _, _ = model(visual[0:1], map_input[0:1], vis_hist[0:1], ego[0:1], 
+        traj_single, _, _ = model(visual[0:1], map_input[0:1], vis_hist[0:1], ego[0:1],
                                   mode="infer")
 
         # Sample 0's output must be identical regardless of what sample 1 contains
@@ -96,114 +85,6 @@ class TestBatchIndependence:
 
 
 # ---------------------------------------------------------------------------
-# 3. View fusion effectiveness — different views must influence output
-# ---------------------------------------------------------------------------
-
-class TestViewFusion:
-    def test_different_views_produce_different_output(self, model, device):
-        """Zeroing one camera view must shift the FUSED feature map.
-
-        Asserts at the fused-feature level rather than the trajectory because
-        in BEV mode the planner's deformable cross-attention samples only a
-        sparse subset of BEV cells — a zeroed view that touches only un-sampled
-        cells could leave the trajectory unchanged in a randomly-initialized
-        model. View fusion's contract is over the fused feature map; that is
-        what this test asserts.
-        """
-        model.eval()
-        torch.manual_seed(42)
-        visual, map_input, vis_hist, ego = make_inputs(1, 7, device)
-        B, V, C, H, W = visual.shape
-
-        def fused_features(x):
-            features = model.Backbone(x.reshape(B * V, C, H, W))
-            return model.FeatureFusion(features, B, V)
-
-        fused_base = fused_features(visual)
-
-        visual_zeroed = visual.clone()
-        visual_zeroed[0, 3] = 0.0
-        fused_zeroed = fused_features(visual_zeroed)
-
-        assert not torch.allclose(fused_base, fused_zeroed, atol=1e-5), \
-            "Zeroing a camera view had no effect on the fused feature map — fusion is broken"
-
-    def test_all_views_contribute(self, model, device):
-        """Each view must influence the FUSED feature map when perturbed.
-
-        View fusion guarantees that every camera view contributes to the
-        fused feature map produced by FeatureFusion. Whether the downstream
-        planner subsequently samples every fused cell is a separate concern
-        — in BEV mode the planner's deformable cross-attention samples only
-        a sparse subset of BEV cells, so a view that touches only un-sampled
-        cells can legitimately have no measurable trajectory influence in a
-        randomly-initialized model. Asserting at the fused-feature level
-        directly tests what view fusion is responsible for and works for
-        all fusion modes.
-        """
-        model.eval()
-        torch.manual_seed(42)
-        visual, map_input, vis_hist, ego = make_inputs(1, 7, device)
-        B, V, C, H, W = visual.shape
-
-        def fused_features(x):
-            features = model.Backbone(x.reshape(B * V, C, H, W))
-            return model.FeatureFusion(features, B, V)
-
-        fused_base = fused_features(visual)
-
-        for view_idx in range(V):
-            visual_mod = visual.clone()
-            visual_mod[0, view_idx] = 5.0
-            fused_mod = fused_features(visual_mod)
-            assert not torch.allclose(fused_base, fused_mod, atol=1e-5), \
-                f"View {view_idx} has no influence on the fused feature map"
-
-    def test_views_contribute_to_fused_with_camera_params(self, build_mock_model, device):
-        """Every view must influence the fused feature map when REAL camera_params
-        are passed through the BEV projection path.
-
-        The other view-contribution tests run only the camera_params=None branch,
-        which exercises the learnable pseudo_projection fallback rather than the
-        geometry-driven projection. Real deployments always pass a [B, V, 3, 4]
-        ego-to-pixel matrix; this test strengthens coverage by feeding one through
-        FeatureFusion and verifying each view still contributes.
-        """
-        model = build_mock_model(num_views=4, fusion_mode="bev", device=device)
-        model.eval()
-        torch.manual_seed(42)
-        visual, map_input, vis_hist, ego = make_inputs(1, 4, device)
-        B, V, C, H, W = visual.shape
-
-        # Identity-like ego-to-pixel projection that places every BEV reference
-        # point inside the image with positive depth on every view:
-        #   u_pix = x_world + 128, v_pix = y_world + 128, depth = 1.
-        # With image_size=256 and the default pc_range
-        # (x in [-60, 120], y in [-60, 60]), normalized image coords land in
-        # [0.27, 0.97] × [0.27, 0.73] — fully visible.
-        cam_params = torch.zeros(B, V, 3, 4, device=device)
-        cam_params[..., 0, 0] = 1.0
-        cam_params[..., 0, 3] = 128.0
-        cam_params[..., 1, 1] = 1.0
-        cam_params[..., 1, 3] = 128.0
-        cam_params[..., 2, 3] = 1.0
-
-        def fused_features(x):
-            features = model.Backbone(x.reshape(B * V, C, H, W))
-            return model.FeatureFusion(features, B, V, camera_params=cam_params)
-
-        fused_base = fused_features(visual)
-
-        for view_idx in range(V):
-            visual_mod = visual.clone()
-            visual_mod[0, view_idx] = 5.0
-            fused_mod = fused_features(visual_mod)
-            assert not torch.allclose(fused_base, fused_mod, atol=1e-5), \
-                f"View {view_idx} has no influence on the fused feature map " \
-                f"under real camera_params projection"
-
-
-# ---------------------------------------------------------------------------
 # 4. Gradient flow — all parameters receive gradients
 # ---------------------------------------------------------------------------
 
@@ -211,7 +92,7 @@ class TestGradientFlow:
     def test_backward_succeeds(self, model, device):
         visual, map_input, vis_hist, ego = make_inputs(2, 7, device)
         target = torch.randn(2, 128, device=device)
-        loss, ego_hidden, future = model(visual, map_input, vis_hist, ego, 
+        loss, ego_hidden, future = model(visual, map_input, vis_hist, ego,
                                          mode="train", trajectory_target=target)
 
         total = loss + ego_hidden.sum() + sum(f.sum() for f in future)
@@ -220,7 +101,7 @@ class TestGradientFlow:
     def test_all_parameters_have_gradients(self, model, device):
         visual, map_input, vis_hist, ego = make_inputs(2, 7, device)
         target = torch.randn(2, 128, device=device)
-        loss, ego_hidden, future = model(visual, map_input, vis_hist, ego, 
+        loss, ego_hidden, future = model(visual, map_input, vis_hist, ego,
                                          mode="train", trajectory_target=target)
 
         total = loss + ego_hidden.sum() + sum(f.sum() for f in future)
@@ -246,7 +127,7 @@ class TestGradientFlow:
         visual, map_input, vis_hist, ego = make_inputs(2, 7, device)
         target = torch.randn(2, 128, device=device)
         loss, ego_hidden, future = model(
-             visual, map_input, vis_hist, ego, 
+             visual, map_input, vis_hist, ego,
             mode="train", trajectory_target=target)
 
         total = loss + ego_hidden.sum() + sum(f.sum() for f in future)
@@ -279,7 +160,7 @@ class TestNumViewsFlexibility:
         visual, map_input, vis_hist, ego = make_inputs(2, num_views, device)
         target = torch.randn(2, 128, device=device)
         loss, ego_hidden, future = model(
-             visual, map_input, vis_hist, ego, 
+             visual, map_input, vis_hist, ego,
             mode="train", trajectory_target=target)
 
         assert loss.dim() == 0
@@ -296,7 +177,7 @@ class TestNumericalStability:
         visual, map_input, vis_hist, ego = make_inputs(2, 7, device)
         target = torch.randn(2, 128, device=device)
         loss, ego_hidden, future = model(
-             visual, map_input, vis_hist, ego, 
+             visual, map_input, vis_hist, ego,
             mode="train", trajectory_target=target)
 
         assert not torch.isnan(loss), "NaN in planner loss"
@@ -308,7 +189,7 @@ class TestNumericalStability:
         visual, map_input, vis_hist, ego = make_inputs(2, 7, device)
         target = torch.randn(2, 128, device=device)
         loss, ego_hidden, future = model(
-             visual, map_input, vis_hist, ego, 
+             visual, map_input, vis_hist, ego,
             mode="train", trajectory_target=target)
 
         assert not torch.isinf(loss), "Inf in planner loss"
@@ -326,584 +207,6 @@ class TestNumericalStability:
 
         assert not torch.isnan(traj).any(), "NaN with large inputs"
         assert not torch.isinf(traj).any(), "Inf with large inputs"
-
-
-# ---------------------------------------------------------------------------
-# Component-level tests
-# ---------------------------------------------------------------------------
-
-class TestFeatureFusionComponent:
-    def test_output_shape(self, device):
-        fusion = FeatureFusion(num_views=8, fusion_mode="concat").to(device)
-        features = [
-            torch.randn(16, 96, 64, 64, device=device),
-            torch.randn(16, 192, 32, 32, device=device),
-            torch.randn(16, 384, 16, 16, device=device),
-            torch.randn(16, 768, 8, 8, device=device),
-        ]
-        out = fusion(features, B=2, V=8)
-        assert out.shape == (2, 256, 8, 8)
-
-    def test_view_reduction_changes_output(self, device):
-        """Verify that view_reduce is not identity (actually mixes views)."""
-        fusion = FeatureFusion(num_views=8, fusion_mode="concat").to(device)
-        fusion.eval()
-
-        features_a = [
-            torch.randn(8, 96, 64, 64, device=device),
-            torch.randn(8, 192, 32, 32, device=device),
-            torch.randn(8, 384, 16, 16, device=device),
-            torch.randn(8, 768, 8, 8, device=device),
-        ]
-        out_a = fusion(features_a, B=1, V=8)
-
-        features_b = [f.clone() for f in features_a]
-        features_b[0][3] = torch.randn_like(features_b[0][3])
-        out_b = fusion(features_b, B=1, V=8)
-
-        assert not torch.allclose(out_a, out_b, atol=1e-5)
-
-
-class TestGRUPlannerComponent:
-    def test_output_shapes(self, device):
-        planner = GRUPlanner(embed_dim=256).to(device)
-        bev = torch.randn(4, 256, 8, 8, device=device)
-        vis_hist = torch.randn(4, 896, device=device)
-        ego = torch.randn(4, 256, device=device)
-        traj, ego_hidden = planner(bev, vis_hist, ego)
-
-        assert traj.shape == (4, 128), "Expected 64 timesteps × 2 signals"
-        assert ego_hidden.shape == (4, 256), "ego_hidden must be 256-dim"
-
-    def test_works_with_arbitrary_bev_resolution(self, device):
-        """Deformable cross-attention via grid_sample should be size-agnostic."""
-        planner = GRUPlanner(embed_dim=256).to(device)
-        vis_hist = torch.randn(2, 896, device=device)
-        ego = torch.randn(2, 256, device=device)
-        for h, w in [(8, 8), (16, 32), (45, 30)]:
-            bev = torch.randn(2, 256, h, w, device=device)
-            traj, ego_hidden = planner(bev, vis_hist, ego)
-            assert traj.shape == (2, 128)
-            assert ego_hidden.shape == (2, 256)
-
-    def test_bev_features_influence_trajectory(self, device):
-        planner = GRUPlanner(embed_dim=256).to(device)
-        planner.eval()
-        vis_hist = torch.randn(1, 896, device=device)
-        ego = torch.randn(1, 256, device=device)
-
-        bev_a = torch.randn(1, 256, 8, 8, device=device)
-        bev_b = torch.randn(1, 256, 8, 8, device=device)
-
-        traj_a, _ = planner(bev_a, vis_hist, ego)
-        traj_b, _ = planner(bev_b, vis_hist, ego)
-
-        assert not torch.allclose(traj_a, traj_b, atol=1e-5), \
-            "Trajectory should depend on BEV features"
-
-    def test_egomotion_influences_trajectory(self, device):
-        planner = GRUPlanner(embed_dim=256).to(device)
-        planner.eval()
-        bev = torch.randn(1, 256, 8, 8, device=device)
-        vis_hist = torch.randn(1, 896, device=device)
-
-        traj_a, _ = planner(bev, vis_hist, torch.randn(1, 256, device=device))
-        traj_b, _ = planner(bev, vis_hist, torch.randn(1, 256, device=device))
-
-        assert not torch.allclose(traj_a, traj_b, atol=1e-5), \
-            "Trajectory should depend on egomotion history"
-
-    def test_visual_history_influences_trajectory(self, device):
-        planner = GRUPlanner(embed_dim=256).to(device)
-        planner.eval()
-        bev = torch.randn(1, 256, 8, 8, device=device)
-        ego = torch.randn(1, 256, device=device)
-
-        traj_a, _ = planner(bev, torch.randn(1, 896, device=device), ego)
-        traj_b, _ = planner(bev, torch.randn(1, 896, device=device), ego)
-
-        assert not torch.allclose(traj_a, traj_b, atol=1e-5), \
-            "Trajectory should depend on visual history"
-
-    def test_configurable_horizon(self, device):
-        planner = GRUPlanner(embed_dim=256, num_timesteps=32, num_signals=3).to(device)
-        bev = torch.randn(2, 256, 8, 8, device=device)
-        vis_hist = torch.randn(2, 896, device=device)
-        ego = torch.randn(2, 256, device=device)
-        traj, _ = planner(bev, vis_hist, ego)
-        assert traj.shape == (2, 32 * 3)
-
-    def test_gradients_flow(self, device):
-        planner = GRUPlanner(embed_dim=256, num_timesteps=4).to(device)
-        bev = torch.randn(1, 256, 8, 8, device=device, requires_grad=True)
-        vis_hist = torch.randn(1, 896, device=device, requires_grad=True)
-        ego = torch.randn(1, 256, device=device, requires_grad=True)
-        traj, ego_hidden = planner(bev, vis_hist, ego)
-        (traj.sum() + ego_hidden.sum()).backward()
-        assert bev.grad is not None and bev.grad.abs().max() > 0
-        assert vis_hist.grad is not None and vis_hist.grad.abs().max() > 0
-        assert ego.grad is not None and ego.grad.abs().max() > 0
-
-    def test_wrong_visual_history_dim_raises(self, device):
-        planner = GRUPlanner(embed_dim=256, visual_history_dim=896).to(device)
-        bev = torch.randn(1, 256, 8, 8, device=device)
-        bad_vis_hist = torch.randn(1, 1024, device=device)  # wrong last dim
-        ego = torch.randn(1, 256, device=device)
-        with pytest.raises(ValueError, match="visual_history last dim must be 896"):
-            planner(bev, bad_vis_hist, ego)
-
-    def test_wrong_egomotion_dim_raises(self, device):
-        planner = GRUPlanner(embed_dim=256, egomotion_dim=256).to(device)
-        bev = torch.randn(1, 256, 8, 8, device=device)
-        vis_hist = torch.randn(1, 896, device=device)
-        bad_ego = torch.randn(1, 128, device=device)  # wrong last dim
-        with pytest.raises(ValueError, match="egomotion_history last dim must be 256"):
-            planner(bev, vis_hist, bad_ego)
-
-    def test_offset_scale_negative_raises(self):
-        with pytest.raises(ValueError, match="offset_scale"):
-            GRUPlanner(embed_dim=256, offset_scale=-1.0)
-
-    def test_offset_scale_nan_raises(self):
-        with pytest.raises(ValueError, match="offset_scale"):
-            GRUPlanner(embed_dim=256, offset_scale=float("nan"))
-
-    def test_offset_scale_inf_raises(self):
-        with pytest.raises(ValueError, match="offset_scale"):
-            GRUPlanner(embed_dim=256, offset_scale=float("inf"))
-
-    def test_offset_scale_bool_raises(self):
-        # bool is an int subclass; the validator must reject it explicitly.
-        with pytest.raises(ValueError, match="offset_scale"):
-            GRUPlanner(embed_dim=256, offset_scale=True)
-
-    def test_offset_scale_non_numeric_raises(self):
-        with pytest.raises(ValueError, match="offset_scale"):
-            GRUPlanner(embed_dim=256, offset_scale="0.1")
-
-    def test_offset_scale_zero_vs_nonzero_differ(self, device):
-        """offset_scale=0 makes deformable attention sample only at the
-        reference point; output must still be valid but differ from the
-        nonzero default."""
-        torch.manual_seed(0)
-        planner_zero = GRUPlanner(embed_dim=256, offset_scale=0.0).to(device)
-        torch.manual_seed(0)
-        planner_pos = GRUPlanner(embed_dim=256, offset_scale=0.1).to(device)
-        planner_zero.eval()
-        planner_pos.eval()
-
-        bev = torch.randn(1, 256, 8, 8, device=device)
-        vis_hist = torch.randn(1, 896, device=device)
-        ego = torch.randn(1, 256, device=device)
-
-        traj_zero, _ = planner_zero(bev, vis_hist, ego)
-        traj_pos, _ = planner_pos(bev, vis_hist, ego)
-
-        assert torch.isfinite(traj_zero).all(), \
-            "offset_scale=0 (reference-point-only) must still produce finite output"
-        assert not torch.allclose(traj_zero, traj_pos, atol=1e-5), \
-            "offset_scale=0 and offset_scale=0.1 should produce different trajectories"
-
-
-class TestFutureStateComponent:
-    def test_accepts_ego_hidden(self, device):
-        future = FutureState(embed_dim=256, ego_hidden_dim=256).to(device)
-        feats = torch.randn(2, 256, 8, 8, device=device)
-        ego_hidden = torch.randn(2, 256, device=device)
-        out = future(feats, ego_hidden)
-        assert len(out) == 4
-        for f in out:
-            assert f.shape == (2, 256, 8, 8)
-
-    def test_ego_hidden_influences_output(self, device):
-        future = FutureState(embed_dim=256, ego_hidden_dim=256).to(device)
-        future.eval()
-        feats = torch.randn(1, 256, 8, 8, device=device)
-
-        out_a = future(feats, torch.randn(1, 256, device=device))
-        out_b = future(feats, torch.randn(1, 256, device=device))
-
-        assert not torch.allclose(out_a[0], out_b[0], atol=1e-5), \
-            "ego_hidden should influence future predictions"
-
-
-# ---------------------------------------------------------------------------
-# Registry tests
-# ---------------------------------------------------------------------------
-
-class TestFusionRegistry:
-    def test_all_modes_registered(self):
-        assert "concat" in FUSION_REGISTRY
-        assert "cross_attn" in FUSION_REGISTRY
-        assert "bev" in FUSION_REGISTRY
-
-    def test_invalid_mode_raises(self):
-        with pytest.raises(ValueError, match="Unknown fusion_mode"):
-            build_view_fusion("nonexistent", num_views=8)
-
-    @pytest.mark.parametrize("fusion_mode", list(FUSION_REGISTRY.keys()))
-    def test_all_modes_produce_correct_shape(self, device, fusion_mode):
-        view_fusion_kwargs = {"bev_h": 8, "bev_w": 8} if fusion_mode == "bev" else {}
-        fusion = FeatureFusion(
-            num_views=8, fusion_mode=fusion_mode,
-            view_fusion_kwargs=view_fusion_kwargs,
-        ).to(device)
-        features = [
-            torch.randn(16, 96, 64, 64, device=device),
-            torch.randn(16, 192, 32, 32, device=device),
-            torch.randn(16, 384, 16, 16, device=device),
-            torch.randn(16, 768, 8, 8, device=device),
-        ]
-        out = fusion(features, B=2, V=8)
-        assert out.shape == (2, 256, 8, 8)
-
-
-# ---------------------------------------------------------------------------
-# Cross-Attention specific tests
-# ---------------------------------------------------------------------------
-
-class TestCrossAttentionFusion:
-    def test_output_shape(self, device):
-        fusion = CrossAttentionViewFusion(num_views=8, embed_dim=256).to(device)
-        x = torch.randn(16, 256, 7, 7, device=device)
-        out = fusion(x, B=2, V=8)
-        assert out.shape == (2, 256, 7, 7)
-
-    def test_view_embeddings_are_learned(self, device):
-        """View embeddings should receive gradients during training."""
-        fusion = CrossAttentionViewFusion(num_views=8, embed_dim=256).to(device)
-        x = torch.randn(8, 256, 7, 7, device=device)
-        out = fusion(x, B=1, V=8)
-        out.sum().backward()
-        assert fusion.view_embed.grad is not None
-        assert fusion.view_embed.grad.abs().max() > 0
-
-    def test_attention_mixes_views(self, device):
-        """Attention should produce different output than simple mean pooling."""
-        fusion = CrossAttentionViewFusion(num_views=4, embed_dim=256).to(device)
-        fusion.eval()
-        x = torch.randn(4, 256, 7, 7, device=device)
-
-        attn_out = fusion(x, B=1, V=4)
-        mean_out = x.reshape(1, 4, 256, 7, 7).mean(dim=1)
-
-        assert not torch.allclose(attn_out, mean_out, atol=1e-3), \
-            "Cross-attention output is identical to naive mean — attention has no effect"
-
-    def test_different_view_orders_produce_different_output(self, device):
-        """Attention with positional embeddings should be order-sensitive."""
-        fusion = CrossAttentionViewFusion(num_views=4, embed_dim=256).to(device)
-        fusion.eval()
-
-        x = torch.randn(4, 256, 7, 7, device=device)
-        out_original = fusion(x, B=1, V=4)
-
-        x_permuted = x[[2, 0, 3, 1]]
-        out_permuted = fusion(x_permuted, B=1, V=4)
-
-        assert not torch.allclose(out_original, out_permuted, atol=1e-5), \
-            "View position embeddings have no effect — output is order-invariant"
-
-
-# ---------------------------------------------------------------------------
-# BEV Fusion specific tests
-# ---------------------------------------------------------------------------
-
-class TestBEVFusion:
-    def test_output_shape(self, device):
-        fusion = BEVViewFusion(num_views=8, embed_dim=256, bev_h=8, bev_w=8).to(device)
-        x = torch.randn(16, 256, 8, 8, device=device)
-        out = fusion(x, B=2, V=8)
-        assert out.shape == (2, 256, 8, 8)
-
-    def test_default_resolution_is_450x300(self):
-        """Production target: 450x300 BEV grid with front-biased pc_range."""
-        fusion = BEVViewFusion(num_views=8, embed_dim=256)
-        assert fusion.bev_h == 450
-        assert fusion.bev_w == 300
-        assert fusion.pc_range == (-60.0, -60.0, -5.0, 120.0, 60.0, 3.0)
-
-    def test_asymmetric_resolution(self, device):
-        """Configurable bev_h != bev_w yields a non-square BEV grid."""
-        fusion = BEVViewFusion(num_views=4, embed_dim=256, bev_h=12, bev_w=20).to(device)
-        x = torch.randn(4, 256, 8, 8, device=device)
-        out = fusion(x, B=1, V=4)
-        assert out.shape == (1, 256, 12, 20)
-
-    def test_output_shape_with_camera_params(self, device):
-        """BEV fusion should work with explicit camera projection matrices."""
-        fusion = BEVViewFusion(num_views=8, embed_dim=256, bev_h=8, bev_w=8).to(device)
-        x = torch.randn(16, 256, 8, 8, device=device)
-        cam_params = torch.randn(2, 8, 3, 4, device=device)
-        out = fusion(x, B=2, V=8, camera_params=cam_params)
-        assert out.shape == (2, 256, 8, 8)
-
-    def test_pseudo_projection_is_learned(self, device):
-        """Without camera_params, pseudo_projection should receive gradients."""
-        fusion = BEVViewFusion(num_views=4, embed_dim=256, bev_h=8, bev_w=8).to(device)
-        x = torch.randn(4, 256, 8, 8, device=device)
-        out = fusion(x, B=1, V=4)
-        out.sum().backward()
-        assert fusion.pseudo_projection.grad is not None
-        assert fusion.pseudo_projection.grad.abs().max() > 0
-
-    def test_bev_queries_are_learned(self, device):
-        """BEV queries should receive gradients during training."""
-        fusion = BEVViewFusion(num_views=4, embed_dim=256, bev_h=8, bev_w=8).to(device)
-        x = torch.randn(4, 256, 8, 8, device=device)
-        out = fusion(x, B=1, V=4)
-        out.sum().backward()
-        assert fusion.bev_queries.weight.grad is not None
-        assert fusion.bev_queries.weight.grad.abs().max() > 0
-
-    def test_camera_params_influence_output(self, device):
-        """Different camera parameters should produce different BEV features."""
-        fusion = BEVViewFusion(num_views=4, embed_dim=256, bev_h=8, bev_w=8).to(device)
-        fusion.eval()
-        x = torch.randn(4, 256, 8, 8, device=device)
-
-        cam_a = torch.randn(1, 4, 3, 4, device=device)
-        cam_b = torch.randn(1, 4, 3, 4, device=device)
-
-        out_a = fusion(x, B=1, V=4, camera_params=cam_a)
-        out_b = fusion(x, B=1, V=4, camera_params=cam_b)
-
-        assert not torch.allclose(out_a, out_b, atol=1e-5), \
-            "Different camera params produced identical output — projection has no effect"
-
-    def test_reference_points_shape(self, device):
-        """3D reference points should have expected shape."""
-        fusion = BEVViewFusion(num_views=8, embed_dim=256, bev_h=7, bev_w=7,
-                               num_points_in_pillar=4).to(device)
-        assert fusion.reference_points_3d.shape == (49, 4, 3)
-
-    def test_no_nan_without_camera_params(self, device):
-        """BEV fusion with pseudo-projection should not produce NaN."""
-        fusion = BEVViewFusion(num_views=8, embed_dim=256, bev_h=8, bev_w=8).to(device)
-        x = torch.randn(16, 256, 8, 8, device=device)
-        out = fusion(x, B=2, V=8)
-        assert not torch.isnan(out).any(), "NaN in BEV output with pseudo-projection"
-
-    def test_points_behind_camera_are_masked(self, device):
-        """Points with negative depth should not contribute to output."""
-        fusion = BEVViewFusion(num_views=1, embed_dim=256, bev_h=8, bev_w=8,
-                               image_size=224,
-                               pc_range=(-10, -10, -5, 10, 10, 3)).to(device)
-
-        # Camera matrix that makes all projected depths negative:
-        # z_proj = row2 @ [x, y, z, 1]^T
-        # Set row2 = [0, 0, -1, -100] so z_proj = -z_world - 100 (always negative
-        # since z_world ranges from -5 to 3 in this pc_range)
-        cam = torch.zeros(1, 1, 3, 4, device=device)
-        cam[0, 0, 0, 0] = 224.0   # fx (irrelevant since depth is negative)
-        cam[0, 0, 1, 1] = 224.0   # fy
-        cam[0, 0, 2, 2] = -1.0    # negate z
-        cam[0, 0, 2, 3] = -100.0  # large negative offset ensures all depths < 0
-
-        ref_2d, mask = fusion._project_to_2d(fusion.reference_points_3d, cam)
-
-        # All points should be masked (behind camera)
-        assert not mask.any(), \
-            "Points behind camera (negative depth) should all be masked"
-
-    def test_projected_center_maps_near_image_center(self, device):
-        """A simple projection should map BEV center to image center."""
-        fusion = BEVViewFusion(num_views=1, embed_dim=256, bev_h=7, bev_w=7,
-                               image_size=224, pc_range=(-1, -1, 0.5, 1, 1, 2)).to(device)
-
-        # Camera: fx=fy=112, cx=cy=112 (image center), z passthrough
-        # BEV center (x=0, y=0) at any z > 0 projects to:
-        #   u = fx*0/z + cx = 112, v = fy*0/z + cy = 112
-        #   normalized: u/224 = 0.5, v/224 = 0.5
-        cam = torch.zeros(1, 1, 3, 4, device=device)
-        cam[0, 0, 0, 0] = 112.0   # fx
-        cam[0, 0, 0, 2] = 112.0   # cx
-        cam[0, 0, 1, 1] = 112.0   # fy
-        cam[0, 0, 1, 2] = 112.0   # cy
-        cam[0, 0, 2, 2] = 1.0     # z passthrough
-
-        ref_2d, mask = fusion._project_to_2d(fusion.reference_points_3d, cam)
-        # ref_2d: [1, 1, 49, num_z, 2]
-
-        # BEV center is query index 24 (7×7 grid, row 3 col 3)
-        center_2d = ref_2d[0, 0, 24, :, :]  # [num_z, 2]
-        center_mask = mask[0, 0, 24, :]      # [num_z]
-
-        # At least some pillar points should be valid
-        assert center_mask.any(), "Center point should have valid projections"
-
-        # Valid points should project exactly to (0.5, 0.5) since x=y=0
-        valid_points = center_2d[center_mask]  # [num_valid, 2]
-        expected = torch.tensor([0.5, 0.5], device=device)
-        assert torch.allclose(valid_points[0], expected, atol=0.01), \
-            f"BEV center should project to image center (0.5, 0.5), got {valid_points[0]}"
-
-    def test_out_of_bounds_points_not_counted_visible(self, device):
-        """When all reference points project out of image bounds, output should be zero."""
-        fusion = BEVViewFusion(num_views=1, embed_dim=256, bev_h=8, bev_w=8,
-                               image_size=224,
-                               pc_range=(-10, -10, -5, 10, 10, 3)).to(device)
-        fusion.eval()
-
-        # Camera that projects everything to far-right of image (u >> image_size)
-        # u = fx * x / z + cx, with fx=1000 and cx=5000, u/224 >> 1 for all points
-        cam = torch.zeros(1, 1, 3, 4, device=device)
-        cam[0, 0, 0, 0] = 1000.0  # fx (very large)
-        cam[0, 0, 0, 2] = 5000.0  # cx (way off image)
-        cam[0, 0, 1, 1] = 1000.0  # fy
-        cam[0, 0, 1, 2] = 5000.0  # cy (way off image)
-        cam[0, 0, 2, 2] = 1.0     # z passthrough (positive depth)
-
-        x = torch.ones(1, 256, 8, 8, device=device)
-        out = fusion(x, B=1, V=1, camera_params=cam)
-
-        # ref_2d normalized = (fx*x/z + cx) / 224 >> 1, so all out of bounds
-        # → mask = False everywhere → visible_count = 0 → has_observation = 0
-        assert out.abs().max() < 1e-6, \
-            "Out-of-bounds projections should produce zero output"
-
-    def test_offset_scale_zero_vs_nonzero_differ(self, device):
-        """offset_scale=0 disables fan-out; output must differ from a nonzero
-        scale at the same seed."""
-        torch.manual_seed(0)
-        fusion_zero = BEVViewFusion(num_views=4, embed_dim=256, bev_h=8, bev_w=8,
-                                    offset_scale=0.0).to(device)
-        torch.manual_seed(0)
-        fusion_pos = BEVViewFusion(num_views=4, embed_dim=256, bev_h=8, bev_w=8,
-                                   offset_scale=0.1).to(device)
-        fusion_zero.eval()
-        fusion_pos.eval()
-        x = torch.randn(4, 256, 8, 8, device=device)
-        out_zero = fusion_zero(x, B=1, V=4)
-        out_pos = fusion_pos(x, B=1, V=4)
-        assert not torch.allclose(out_zero, out_pos, atol=1e-5), \
-            "offset_scale=0 and offset_scale=0.1 produced identical BEV output"
-
-    def test_offset_scale_negative_raises(self):
-        with pytest.raises(ValueError, match="offset_scale"):
-            BEVViewFusion(num_views=4, embed_dim=256, bev_h=8, bev_w=8,
-                          offset_scale=-1.0)
-
-    def test_offset_scale_nan_raises(self):
-        with pytest.raises(ValueError, match="offset_scale"):
-            BEVViewFusion(num_views=4, embed_dim=256, bev_h=8, bev_w=8,
-                          offset_scale=float("nan"))
-
-    def test_offset_scale_inf_raises(self):
-        with pytest.raises(ValueError, match="offset_scale"):
-            BEVViewFusion(num_views=4, embed_dim=256, bev_h=8, bev_w=8,
-                          offset_scale=float("inf"))
-
-    def test_offset_scale_bool_raises(self):
-        # bool is an int subclass; the validator must reject it explicitly.
-        with pytest.raises(ValueError, match="offset_scale"):
-            BEVViewFusion(num_views=4, embed_dim=256, bev_h=8, bev_w=8,
-                          offset_scale=True)
-
-    def test_offset_scale_non_numeric_raises(self):
-        with pytest.raises(ValueError, match="offset_scale"):
-            BEVViewFusion(num_views=4, embed_dim=256, bev_h=8, bev_w=8,
-                          offset_scale="0.1")
-
-    def test_no_visible_camera_produces_zero_output(self, device):
-        """If no camera can see any BEV cell, output should be exactly zero."""
-        fusion = BEVViewFusion(num_views=1, embed_dim=256, bev_h=8, bev_w=8,
-                               image_size=224,
-                               pc_range=(-10, -10, -5, 10, 10, 3)).to(device)
-        fusion.eval()
-
-        x = torch.ones(1, 256, 8, 8, device=device)
-
-        # Camera that places everything behind (negative depth)
-        cam_behind = torch.zeros(1, 1, 3, 4, device=device)
-        cam_behind[0, 0, 2, 2] = -1.0
-        cam_behind[0, 0, 2, 3] = -100.0
-        out = fusion(x, B=1, V=1, camera_params=cam_behind)
-
-        # has_observation mask zeroes output after FFN
-        assert out.abs().max() < 1e-6, \
-            "No visible camera should produce zero BEV features"
-
-
-
-# ---------------------------------------------------------------------------
-# Integration tests — full backbone (slow, marked for separate CI tier)
-# ---------------------------------------------------------------------------
-
-@pytest.mark.integration
-class TestFullBackboneIntegration:
-    """End-to-end tests with the real pretrained backbone.
-
-    These verify that the full pipeline (backbone → fusion → planner → future)
-    produces correct shapes and numerically stable outputs. Run separately
-    from unit tests via: pytest -m integration
-    """
-
-    def test_full_forward_pass(self, full_model, device):
-        """Smoke test: full model forward produces expected output shapes."""
-        visual, map_input, vis_hist, ego = make_inputs(1, 7, device)
-        target = torch.randn(1, 128, device=device)
-        loss, ego_hidden, future = full_model(
-             visual, map_input, vis_hist, ego, 
-            mode="train", trajectory_target=target)
-
-        assert loss.dim() == 0
-        assert ego_hidden.shape == (1, 256)
-        assert len(future) == 4
-        for f in future:
-            assert f.shape == (1, 256, 8, 8)
-
-        traj, _, _ = full_model(visual, vis_hist, ego, mode="infer")
-        assert traj.shape == (1, 128)
-
-    def test_full_forward_no_nan(self, full_model, device):
-        """Full pipeline must not produce NaN with real backbone weights."""
-        visual, map_input, vis_hist, ego = make_inputs(2, 7, device)
-        target = torch.randn(2, 128, device=device)
-        loss, ego_hidden, future = full_model(visual, map_input, vis_hist, ego,
-                                              mode="train", trajectory_target=target)
-
-        assert not torch.isnan(loss)
-        assert not torch.isnan(ego_hidden).any()
-        for f in future:
-            assert not torch.isnan(f).any()
-
-
-@pytest.mark.integration
-class TestResNet50Backbone:
-    """Exercises the dynamic backbone_channels computation on a backbone
-    whose feature_info shape differs from Swin (5 stages of channels
-    64/256/512/1024/2048 vs Swin's 4 stages of 96/192/384/768)."""
-
-    def test_resnet50_forward_pass(self, device):
-        from model_components.auto_e2e import AutoE2E
-        try:
-            model = AutoE2E(
-                backbone="res_net_50", num_views=7, fusion_mode="concat",
-                is_pretrained=False,
-            ).to(device)
-        except (FileNotFoundError, OSError) as e:
-            pytest.skip(f"Backbone construction failed: {e}")
-
-        # Dynamic backbone_channels = sum of all 5 ResNet50 stages = 3904
-        assert model.Backbone.backbone_channels == 64 + 256 + 512 + 1024 + 2048
-
-        visual, map_input, vis_hist, ego = make_inputs(1, 7, device)
-        target = torch.randn(1, 128, device=device)
-        loss, ego_hidden, future = model(
-            visual, map_input, vis_hist, ego, mode="train", trajectory_target=target)
-
-        assert loss.dim() == 0
-        assert ego_hidden.shape == (1, 256)
-        assert len(future) == 4
-        for f in future:
-            assert f.shape == (1, 256, 8, 8)
-        assert torch.isfinite(loss)
-        assert torch.isfinite(ego_hidden).all()
-
-        traj, _, _ = model(visual, vis_hist, ego, mode="infer")
-        assert traj.shape == (1, 128)
-        assert torch.isfinite(traj).all()
 
 
 # ---------------------------------------------------------------------------
@@ -1083,40 +386,6 @@ class TestTrajectoryDynamics:
             "offset_scale=0 vs 0.1 produced identical attn output — offsets inactive"
 
 
-class TestFutureStateChunkSplit:
-    def test_four_outputs_are_distinct(self, device):
-        """torch.chunk must split along channels, not return 4 views of the same data."""
-        torch.manual_seed(0)
-        future = FutureState(embed_dim=256, ego_hidden_dim=256).to(device)
-        future.eval()
-        feats = torch.randn(2, 256, 8, 8, device=device)
-        ego_hidden = torch.randn(2, 256, device=device)
-
-        out = future(feats, ego_hidden)
-        assert len(out) == 4
-        for i in range(4):
-            for j in range(i + 1, 4):
-                assert not torch.allclose(out[i], out[j], atol=1e-5), \
-                    f"FutureState outputs {i} and {j} are identical — chunk is broken"
-
-    def test_ego_hidden_changes_all_four_outputs(self, device):
-        """Different ego_hidden must shift every one of the 4 future predictions."""
-        torch.manual_seed(0)
-        future = FutureState(embed_dim=256, ego_hidden_dim=256).to(device)
-        future.eval()
-        feats = torch.randn(1, 256, 8, 8, device=device)
-
-        ego_a = torch.randn(1, 256, device=device)
-        ego_b = torch.randn(1, 256, device=device)
-
-        out_a = future(feats, ego_a)
-        out_b = future(feats, ego_b)
-
-        for i in range(4):
-            assert not torch.allclose(out_a[i], out_b[i], atol=1e-5), \
-                f"Future output {i} did not change when ego_hidden changed"
-
-
 class TestVisualHistoryNonZeroDifference:
     def test_two_nonzero_visual_histories_differ(self, device):
         """Both visual_history inputs are non-zero and distinct — outputs must differ."""
@@ -1189,532 +458,3 @@ class TestFullPipelineRobustness:
         assert ego_hidden.shape == (1, 256)
         assert torch.isfinite(traj).all()
         assert torch.isfinite(ego_hidden).all()
-
-
-class _StubBackboneWithFeatureInfo(torch.nn.Module):
-    """Channels-first backbone exposing timm-style feature_info."""
-
-    def __init__(self):
-        super().__init__()
-        self.stage0 = torch.nn.Conv2d(3, 32, 3, stride=2, padding=1)
-        self.stage1 = torch.nn.Conv2d(32, 48, 3, stride=2, padding=1)
-        self.stage2 = torch.nn.Conv2d(48, 64, 3, stride=2, padding=1)
-        self.feature_info = [{"num_chs": 32}, {"num_chs": 48}, {"num_chs": 64}]
-
-    def forward(self, x):
-        s0 = self.stage0(x)
-        s1 = self.stage1(s0)
-        s2 = self.stage2(s1)
-        return [s0, s1, s2]
-
-
-class _StubBackboneNoFeatureInfo(torch.nn.Module):
-    """Channels-first backbone with NO feature_info (probe fallback path)."""
-
-    def __init__(self):
-        super().__init__()
-        self.stage0 = torch.nn.Conv2d(3, 24, 3, stride=2, padding=1)
-        self.stage1 = torch.nn.Conv2d(24, 56, 3, stride=2, padding=1)
-        self.stage2 = torch.nn.Conv2d(56, 112, 3, stride=2, padding=1)
-
-    def forward(self, x):
-        s0 = self.stage0(x)
-        s1 = self.stage1(s0)
-        s2 = self.stage2(s1)
-        return [s0, s1, s2]
-
-
-class _StubBackboneSwinLike(torch.nn.Module):
-    """Channels-last backbone (B, H, W, C) — exercises permute branch."""
-
-    def __init__(self):
-        super().__init__()
-        self.stage0 = torch.nn.Conv2d(3, 32, 3, stride=2, padding=1)
-        self.stage1 = torch.nn.Conv2d(32, 48, 3, stride=2, padding=1)
-        self.feature_info = [{"num_chs": 32}, {"num_chs": 48}]
-
-    def forward(self, x):
-        s0_cf = self.stage0(x)                                  # [B, 32, H, W]
-        s1_cf = self.stage1(s0_cf)                              # [B, 48, H, W]
-        s0 = s0_cf.permute(0, 2, 3, 1).contiguous()             # [B, H, W, 32]
-        s1 = s1_cf.permute(0, 2, 3, 1).contiguous()             # [B, H, W, 48]
-        return [s0, s1]
-
-
-class TestBackboneChannelDiscovery:
-    """Cover the backbone_channels discovery + layout-detection in Backbone."""
-
-    def _make_backbone(self, monkeypatch, stub_module):
-        # Patch the registry call so build_backbone returns our stub.
-        monkeypatch.setattr(
-            "model_components.backbone.build_backbone",
-            lambda *a, **kw: stub_module,
-        )
-        return Backbone(backbone="stub", is_pretrained=False)
-
-    def test_feature_info_path_sums_channels(self, monkeypatch):
-        bb = self._make_backbone(monkeypatch, _StubBackboneWithFeatureInfo())
-        assert bb.backbone_channels == 32 + 48 + 64
-
-    def test_probe_fallback_when_feature_info_missing(self, monkeypatch):
-        bb = self._make_backbone(monkeypatch, _StubBackboneNoFeatureInfo())
-        # No feature_info — channels recovered via probing.
-        assert bb.backbone_channels == 24 + 56 + 112
-
-    def test_feature_info_channels_match_forward_output(self, monkeypatch, device):
-        """sum(feature_info channels) must equal the actual concat-channel dim
-        of the forward output."""
-        bb = self._make_backbone(monkeypatch, _StubBackboneWithFeatureInfo()).to(device)
-        x = torch.randn(2, 3, 32, 32, device=device)
-        feats = bb(x)
-        total_c = sum(f.shape[1] for f in feats)
-        assert total_c == bb.backbone_channels
-
-    def test_probe_channels_match_forward_output(self, monkeypatch, device):
-        bb = self._make_backbone(monkeypatch, _StubBackboneNoFeatureInfo()).to(device)
-        x = torch.randn(2, 3, 32, 32, device=device)
-        feats = bb(x)
-        total_c = sum(f.shape[1] for f in feats)
-        assert total_c == bb.backbone_channels
-
-    def test_channels_last_backbone_is_permuted(self, monkeypatch, device):
-        """Channels-last (B, H, W, C) output must be permuted to (B, C, H, W)
-        based on tensor shape, NOT on the backbone name."""
-        bb = self._make_backbone(monkeypatch, _StubBackboneSwinLike()).to(device)
-        x = torch.randn(2, 3, 32, 32, device=device)
-        feats = bb(x)
-        # After Backbone.forward, every feature must be channels-first with the
-        # expected channel count at dim 1.
-        assert feats[0].shape[1] == 32
-        assert feats[1].shape[1] == 48
-
-    def test_channels_first_backbone_not_permuted(self, monkeypatch, device):
-        bb = self._make_backbone(monkeypatch, _StubBackboneWithFeatureInfo()).to(device)
-        x = torch.randn(2, 3, 32, 32, device=device)
-        feats = bb(x)
-        for f, expected in zip(feats, [32, 48, 64]):
-            assert f.shape[1] == expected
-
-
-class TestFeatureFusionWithSwinChannels:
-    def test_dynamic_backbone_channels_with_swin_sizes(self, device):
-        """FeatureFusion should accept Swin's per-stage channels (96, 192, 384, 768)
-        at their natural spatial resolutions and produce the expected fused shape."""
-        backbone_channels = 96 + 192 + 384 + 768  # 1440
-        fusion = FeatureFusion(
-            num_views=8, backbone_channels=backbone_channels, fusion_mode="concat",
-        ).to(device)
-
-        # Per-stage Swin spatial dims for a 256x256 input
-        features = [
-            torch.randn(16, 96, 64, 64, device=device),
-            torch.randn(16, 192, 32, 32, device=device),
-            torch.randn(16, 384, 16, 16, device=device),
-            torch.randn(16, 768, 8, 8, device=device),
-        ]
-        out = fusion(features, B=2, V=8)
-        assert out.shape == (2, 256, 8, 8)
-        assert torch.isfinite(out).all()
-
-
-# ---------------------------------------------------------------------------
-# Planner registry / Flow Matching planner / backcompat
-# ---------------------------------------------------------------------------
-
-
-class TestPlannerRegistry:
-    def test_all_modes_registered(self):
-        assert "gru" in PLANNER_REGISTRY
-        assert "flow_matching" in PLANNER_REGISTRY
-
-    def test_invalid_mode_raises(self):
-        with pytest.raises(ValueError, match="Unknown planner_mode"):
-            build_planner("nonexistent", embed_dim=256)
-
-    def test_build_returns_correct_type(self):
-        gru = build_planner("gru", embed_dim=256)
-        assert isinstance(gru, GRUPlanner)
-        fm = build_planner("flow_matching", embed_dim=256)
-        assert isinstance(fm, FlowMatchingPlanner)
-
-
-class TestFlowMatchingPlanner:
-    def test_construct_training_data_shapes(self, device):
-        planner = FlowMatchingPlanner(embed_dim=256).to(device)
-        target = torch.randn(4, 128, device=device)
-        u_t, t, target_velocity = planner.construct_training_data(target)
-        assert u_t.shape == (4, 128)
-        assert t.shape == (4,)
-        assert target_velocity.shape == (4, 128)
-        assert (t >= 0).all() and (t <= 1).all()
-
-    def test_compute_planner_loss_end_to_end(self, device):
-        """The canonical training-loop pattern must work:
-        compute_planner_loss returns a scalar loss + ego_hidden, and
-        backprop reaches the BEV input."""
-        planner = FlowMatchingPlanner(embed_dim=256).to(device)
-        planner.train()
-        bev = torch.randn(2, 256, 8, 8, device=device, requires_grad=True)
-        vis_hist = torch.randn(2, 896, device=device)
-        ego = torch.randn(2, 256, device=device)
-        target = torch.randn(2, 128, device=device)
-
-        loss, ego_hidden = planner.compute_planner_loss(
-            bev, vis_hist, ego, target,
-        )
-        assert loss.dim() == 0
-        assert ego_hidden.shape == (2, 256)
-        assert torch.isfinite(loss)
-
-        loss.backward()
-        assert bev.grad is not None and bev.grad.abs().max() > 0
-
-    def test_inference_forward_returns_trajectory_shape(self, device):
-        planner = FlowMatchingPlanner(embed_dim=256).to(device)
-        planner.eval()
-        bev = torch.randn(2, 256, 8, 8, device=device)
-        vis_hist = torch.randn(2, 896, device=device)
-        ego = torch.randn(2, 256, device=device)
-        traj, ego_hidden = planner(bev, vis_hist, ego)
-        assert traj.shape == (2, 128)
-        assert ego_hidden.shape == (2, 256)
-
-    def test_inference_output_is_finite(self, device):
-        planner = FlowMatchingPlanner(embed_dim=256, num_inference_steps=10).to(device)
-        planner.eval()
-        bev = torch.randn(1, 256, 8, 8, device=device)
-        vis_hist = torch.randn(1, 896, device=device)
-        ego = torch.randn(1, 256, device=device)
-        traj, _ = planner(bev, vis_hist, ego)
-        assert torch.isfinite(traj).all()
-
-    def _v_theta(self, planner, bev, vis_hist, ego, u_t, t):
-        """Run the velocity network at fixed (u_t, t) — bypasses the
-        public API so tests can pin all three inputs."""
-        mod_cond = planner._modulation_conditioning(vis_hist, ego)
-        bev_seq = planner._project_bev(bev)
-        return planner._v_theta(u_t, t, bev_seq, mod_cond)
-
-    def test_velocity_depends_on_bev(self, device):
-        torch.manual_seed(0)
-        planner = FlowMatchingPlanner(embed_dim=256).to(device)
-        planner.eval()
-        vis_hist = torch.randn(1, 896, device=device)
-        ego = torch.randn(1, 256, device=device)
-        u_t = torch.randn(1, 128, device=device)
-        t = torch.tensor([0.5], device=device)
-
-        bev_a = torch.randn(1, 256, 8, 8, device=device)
-        bev_b = torch.randn(1, 256, 8, 8, device=device)
-
-        v_a = self._v_theta(planner, bev_a, vis_hist, ego, u_t, t)
-        v_b = self._v_theta(planner, bev_b, vis_hist, ego, u_t, t)
-        assert not torch.allclose(v_a, v_b, atol=1e-5), \
-            "v_theta is not sensitive to BEV features"
-
-    def test_velocity_depends_on_timestep(self, device):
-        torch.manual_seed(0)
-        planner = FlowMatchingPlanner(embed_dim=256).to(device)
-        planner.eval()
-        bev = torch.randn(1, 256, 8, 8, device=device)
-        vis_hist = torch.randn(1, 896, device=device)
-        ego = torch.randn(1, 256, device=device)
-        u_t = torch.randn(1, 128, device=device)
-
-        v_t1 = self._v_theta(planner, bev, vis_hist, ego, u_t,
-                             torch.tensor([0.1], device=device))
-        v_t2 = self._v_theta(planner, bev, vis_hist, ego, u_t,
-                             torch.tensor([0.9], device=device))
-        assert not torch.allclose(v_t1, v_t2, atol=1e-5), \
-            "v_theta is not sensitive to flow timestep"
-
-    def test_velocity_depends_on_conditioning(self, device):
-        torch.manual_seed(0)
-        planner = FlowMatchingPlanner(embed_dim=256).to(device)
-        planner.eval()
-        bev = torch.randn(1, 256, 8, 8, device=device)
-        u_t = torch.randn(1, 128, device=device)
-        t = torch.tensor([0.5], device=device)
-
-        v_a = self._v_theta(
-            planner, bev,
-            torch.randn(1, 896, device=device),
-            torch.randn(1, 256, device=device),
-            u_t, t,
-        )
-        v_b = self._v_theta(
-            planner, bev,
-            torch.randn(1, 896, device=device),
-            torch.randn(1, 256, device=device),
-            u_t, t,
-        )
-        assert not torch.allclose(v_a, v_b, atol=1e-5), \
-            "v_theta is not sensitive to ego/visual_history conditioning"
-
-    def test_inference_differs_from_noise(self, device):
-        """Euler integration must actually transform the noise — output
-        cannot match the input noise sample."""
-        torch.manual_seed(0)
-        planner = FlowMatchingPlanner(embed_dim=256, num_inference_steps=10).to(device)
-        planner.eval()
-        bev = torch.randn(1, 256, 8, 8, device=device)
-        vis_hist = torch.randn(1, 896, device=device)
-        ego = torch.randn(1, 256, device=device)
-
-        torch.manual_seed(123)
-        x0 = torch.randn(1, 128, device=device)
-        torch.manual_seed(123)  # same seed — planner draws an identical x0 inside
-        traj, _ = planner(bev, vis_hist, ego)
-        assert not torch.allclose(traj, x0, atol=1e-3), \
-            "Inference output equals the input noise — ODE did not advance"
-
-    def test_gradient_flows(self, device):
-        planner = FlowMatchingPlanner(embed_dim=256).to(device)
-        bev = torch.randn(1, 256, 8, 8, device=device, requires_grad=True)
-        vis_hist = torch.randn(1, 896, device=device, requires_grad=True)
-        ego = torch.randn(1, 256, device=device, requires_grad=True)
-        target = torch.randn(1, 128, device=device)
-        loss, ego_hidden = planner.compute_planner_loss(
-            bev, vis_hist, ego, target,
-        )
-        (loss + ego_hidden.sum()).backward()
-        assert bev.grad is not None and bev.grad.abs().max() > 0
-        assert vis_hist.grad is not None and vis_hist.grad.abs().max() > 0
-        assert ego.grad is not None and ego.grad.abs().max() > 0
-
-    def test_inference_generator_is_reproducible(self, device):
-        """A shared torch.Generator must make inference deterministic across
-        runs, and different seeds must produce different trajectories."""
-        planner = FlowMatchingPlanner(embed_dim=256, num_inference_steps=10).to(device)
-        planner.eval()
-        bev = torch.randn(2, 256, 8, 8, device=device)
-        vis_hist = torch.randn(2, 896, device=device)
-        ego = torch.randn(2, 256, device=device)
-
-        gen_a = torch.Generator(device=device).manual_seed(42)
-        gen_b = torch.Generator(device=device).manual_seed(42)
-        gen_c = torch.Generator(device=device).manual_seed(7)
-
-        traj_a, _ = planner(bev, vis_hist, ego, generator=gen_a)
-        traj_b, _ = planner(bev, vis_hist, ego, generator=gen_b)
-        traj_c, _ = planner(bev, vis_hist, ego, generator=gen_c)
-
-        assert torch.equal(traj_a, traj_b), \
-            "same generator seed must produce identical inference trajectories"
-        assert not torch.allclose(traj_a, traj_c), \
-            "different generator seeds must produce different trajectories"
-
-    def test_compute_planner_loss_wrong_target_shape_raises(self, device):
-        planner = FlowMatchingPlanner(embed_dim=256).to(device)
-        bev = torch.randn(2, 256, 8, 8, device=device)
-        vis_hist = torch.randn(2, 896, device=device)
-        ego = torch.randn(2, 256, device=device)
-        # Wrong batch dim
-        bad_target = torch.randn(3, 128, device=device)
-        with pytest.raises(ValueError, match="trajectory_target must have shape"):
-            planner.compute_planner_loss(bev, vis_hist, ego, bad_target)
-        # Wrong feature dim
-        bad_target2 = torch.randn(2, 64, device=device)
-        with pytest.raises(ValueError, match="trajectory_target must have shape"):
-            planner.compute_planner_loss(bev, vis_hist, ego, bad_target2)
-
-    def test_construct_training_data_wrong_target_shape_propagates(self, device):
-        """The internal _validate_flow_inputs guard must catch shape regressions
-        even when the user only calls construct_training_data."""
-        planner = FlowMatchingPlanner(
-            embed_dim=256, num_timesteps=4, num_signals=2,
-        ).to(device)
-        # construct_training_data uses target's shape verbatim; if that shape
-        # disagrees with planner.trajectory_dim, the internal validator fires.
-        bad_target = torch.randn(2, 16, device=device)  # 16 != 4*2
-        with pytest.raises(ValueError, match="noisy_trajectory must have shape"):
-            planner.construct_training_data(bad_target)
-
-    def test_validate_flow_inputs_shape_dtype(self, device):
-        planner = FlowMatchingPlanner(embed_dim=256).to(device)
-        good_u = torch.randn(2, 128, device=device)
-        good_t = torch.rand(2, device=device)
-        # Wrong u_t shape
-        with pytest.raises(ValueError, match="noisy_trajectory must have shape"):
-            planner._validate_flow_inputs(torch.randn(2, 64, device=device),
-                                          good_t, batch_size=2)
-        # Wrong t shape
-        with pytest.raises(ValueError, match="flow_timestep must have shape"):
-            planner._validate_flow_inputs(good_u,
-                                          torch.rand(3, device=device),
-                                          batch_size=2)
-        # Dtype mismatch
-        good_u_f64 = good_u.to(torch.float64)
-        with pytest.raises(ValueError, match="must share dtype"):
-            planner._validate_flow_inputs(good_u_f64, good_t, batch_size=2)
-
-    @pytest.mark.skipif(not torch.cuda.is_available(),
-                        reason="device-mismatch case requires CUDA")
-    def test_validate_flow_inputs_device_mismatch(self):
-        planner = FlowMatchingPlanner(embed_dim=256)
-        u_cpu = torch.randn(1, 128)
-        t_cuda = torch.rand(1, device="cuda")
-        with pytest.raises(ValueError, match="must be on the same device"):
-            planner._validate_flow_inputs(u_cpu, t_cuda, batch_size=1)
-
-    def test_timestep_sampler_beta_in_range(self, device):
-        """Default shifted-Beta sampler stays inside [0, beta_scale] and
-        matches the trajectory_target dtype."""
-        torch.manual_seed(0)
-        planner = FlowMatchingPlanner(embed_dim=256).to(device)
-        target = torch.randn(256, 128, device=device)
-        _, t, _ = planner.construct_training_data(target)
-        assert t.shape == (256,)
-        assert t.dtype == target.dtype
-        assert (t >= 0).all()
-        assert (t <= planner.beta_scale + 1e-6).all()
-
-    def test_timestep_sampler_beta_biased_toward_low_t(self, device):
-        """Documents the noisy-end bias: shifted Beta(1.5, 1) puts mass
-        below 0.5 in expectation. Use a loose bound so the assertion is
-        not flaky."""
-        torch.manual_seed(0)
-        planner = FlowMatchingPlanner(embed_dim=256).to(device)
-        target = torch.randn(2048, 128, device=device)
-        _, t, _ = planner.construct_training_data(target)
-        assert t.mean().item() < 0.5
-
-    def test_timestep_sampler_uniform_option(self, device):
-        """timestep_sampler='uniform' recovers U(0, 1)."""
-        planner = FlowMatchingPlanner(
-            embed_dim=256, timestep_sampler="uniform",
-        ).to(device)
-        target = torch.randn(256, 128, device=device)
-        _, t, _ = planner.construct_training_data(target)
-        assert t.shape == (256,)
-        assert t.dtype == target.dtype
-        assert (t >= 0).all()
-        assert (t < 1).all()
-
-    def test_invalid_timestep_sampler_raises(self):
-        with pytest.raises(ValueError, match="timestep_sampler"):
-            FlowMatchingPlanner(timestep_sampler="gaussian")
-        with pytest.raises(ValueError, match="beta_alpha"):
-            FlowMatchingPlanner(beta_alpha=0)
-        with pytest.raises(ValueError, match="beta_scale"):
-            FlowMatchingPlanner(beta_scale=1.5)
-
-
-class TestGRUPlannerBackcompat:
-    """The GRU planner moved into the trajectory_planning subpackage —
-    its public behavior must be unchanged."""
-
-    def test_shape_unchanged(self, device):
-        planner = GRUPlanner(embed_dim=256).to(device)
-        bev = torch.randn(2, 256, 8, 8, device=device)
-        vis_hist = torch.randn(2, 896, device=device)
-        ego = torch.randn(2, 256, device=device)
-        traj, ego_hidden = planner(bev, vis_hist, ego)
-        assert traj.shape == (2, 128)
-        assert ego_hidden.shape == (2, 256)
-
-    def test_mode_argument_accepted(self, device):
-        """forward must accept mode without changing GRU output."""
-        torch.manual_seed(0)
-        planner = GRUPlanner(embed_dim=256).to(device)
-        planner.eval()
-        bev = torch.randn(1, 256, 8, 8, device=device)
-        vis_hist = torch.randn(1, 896, device=device)
-        ego = torch.randn(1, 256, device=device)
-        traj_train, _ = planner(bev, vis_hist, ego, mode="train")
-        traj_infer, _ = planner(bev, vis_hist, ego, mode="infer")
-        assert torch.allclose(traj_train, traj_infer)
-
-    def test_extra_kwargs_ignored(self, device):
-        """forward must silently swallow flow-matching-only kwargs."""
-        planner = GRUPlanner(embed_dim=256).to(device)
-        planner.eval()
-        bev = torch.randn(1, 256, 8, 8, device=device)
-        vis_hist = torch.randn(1, 896, device=device)
-        ego = torch.randn(1, 256, device=device)
-        traj, _ = planner(
-            bev, vis_hist, ego,
-            trajectory_target=torch.randn(1, 128, device=device),
-            noisy_trajectory=torch.randn(1, 128, device=device),
-            flow_timestep=torch.tensor([0.5], device=device),
-        )
-        assert traj.shape == (1, 128)
-
-
-class TestAutoE2EWithFlowMatching:
-    @staticmethod
-    def _fm_model(build_mock_model, device):
-        return build_mock_model(
-            num_views=8, fusion_mode="bev", device=device,
-            planner_mode="flow_matching",
-            planner_kwargs={"num_inference_steps": 4},
-        )
-
-    def test_train_mode_returns_scalar_loss(self, build_mock_model, device):
-        """Under the uniform Option-B contract, train mode must return a
-        scalar planner loss — NOT the raw flow-matching velocity tensor.
-        This documents that the velocity-vs-target footgun is gone."""
-        model = self._fm_model(build_mock_model, device)
-        model.train()
-        visual, map_input, vis_hist, ego = make_inputs(2, 8, device)
-        target = torch.randn(2, 128, device=device)
-        loss, ego_hidden, future = model(
-            visual, map_input, vis_hist, ego, mode="train", trajectory_target=target,
-        )
-        assert loss.dim() == 0, \
-            "FM train mode must expose a scalar loss, not a [B, T*S] velocity"
-        assert loss.requires_grad
-        assert ego_hidden.shape == (2, 256)
-        assert future is not None and len(future) == 4
-
-    def test_infer_mode_returns_trajectory(self, build_mock_model, device):
-        model = self._fm_model(build_mock_model, device)
-        model.eval()
-        visual, map_input, vis_hist, ego = make_inputs(1, 8, device)
-        traj, ego_hidden, future = model(visual, map_input, vis_hist, ego, mode="infer")
-        assert traj.shape == (1, 128)
-        assert ego_hidden.shape == (1, 256)
-        assert future is None
-        assert torch.isfinite(traj).all()
-
-    def test_backward_flows_through_planner_loss(self, build_mock_model, device):
-        model = self._fm_model(build_mock_model, device)
-        model.train()
-        visual, map_input, vis_hist, ego = make_inputs(2, 8, device)
-        target = torch.randn(2, 128, device=device)
-        loss, ego_hidden, future = model(
-            visual, map_input, vis_hist, ego, mode="train", trajectory_target=target,
-        )
-        total = loss + ego_hidden.sum() + sum(f.sum() for f in future)
-        total.backward()
-        # At least one Backbone and one TrajectoryPlanner param must see grad.
-        backbone_grad = any(
-            p.grad is not None and p.grad.abs().max() > 0
-            for n, p in model.named_parameters() if n.startswith("Backbone.")
-        )
-        planner_grad = any(
-            p.grad is not None and p.grad.abs().max() > 0
-            for n, p in model.named_parameters()
-            if n.startswith("TrajectoryPlanner.")
-        )
-        assert backbone_grad and planner_grad
-
-    def test_train_mode_requires_target(self, build_mock_model, device):
-        model = self._fm_model(build_mock_model, device)
-        visual, map_input, vis_hist, ego = make_inputs(1, 8, device)
-        with pytest.raises(ValueError, match="trajectory_target"):
-            model(visual, map_input, vis_hist, ego, mode="train")
-
-    def test_gru_and_fm_interchangeable_at_inference(self, build_mock_model, device):
-        """GRU and FM both produce a [B, T*S] trajectory in inference mode —
-        callers can swap planners with no other code changes."""
-        gru_model = build_mock_model(num_views=8, fusion_mode="concat",
-                                     device=device)
-        fm_model = self._fm_model(build_mock_model, device)
-        gru_model.eval()
-        fm_model.eval()
-
-        visual, map_input, vis_hist, ego = make_inputs(2, 8, device)
-        gru_traj, _, _ = gru_model(visual, map_input, vis_hist, ego, mode="infer")
-        fm_traj, _, _ = fm_model(visual, map_input, vis_hist, ego, mode="infer")
-        assert gru_traj.shape == fm_traj.shape == (2, 128)
